add Permissions admin_service
Fix: test_delete_all_users

+++ src/services/rest/admin_service.py
@@ -1,3 +1,4 @@
+from src.data.role import Role
 from src.services.rest.main_service import MainService
 from http import HTTPStatus
 import pymongo
@@ -16,26 +17,39 @@ class AdminService(MainService):
         self.objects = super().get_db().objects
         self.commands = super().get_db().commands
 
-    def delete_users(self) -> tuple:
+    def delete_users(self, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
+
         self.users.delete_many({})
         return '', HTTPStatus.NO_CONTENT
 
-    def get_all_users(self, page: int, limit: int) -> tuple:
+    def get_all_users(self, page: int, limit: int, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
         to_skip = calc_to_skip(page, limit)
         return list(self.users.find({}).sort("name", ASCENDING).skip(to_skip).limit(limit)), HTTPStatus.OK
 
-    def delete_objects(self) -> tuple:
+    def delete_objects(self, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
         self.objects.delete_many({})
         return '', HTTPStatus.NO_CONTENT
 
-    def get_all_objects(self, page: int, limit: int) -> tuple:
+    def get_all_objects(self, page: int, limit: int, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
         to_skip = calc_to_skip(page, limit)
         return list(self.objects.find({}).sort("type", ASCENDING).skip(to_skip).limit(limit)), HTTPStatus.OK
 
-    def delete_commands(self) -> tuple:
+    def delete_commands(self, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
         self.commands.delete_many({})
         return '', HTTPStatus.NO_CONTENT
 
-    def get_all_commands(self, page: int, limit: int) -> tuple:
+    def get_all_commands(self, page: int, limit: int, email: str) -> tuple:
+        if not super().check_permissions(Role.ADMIN, email):
+            return {"Error": "User doesn't have permissions"}, HTTPStatus.UNAUTHORIZED
         to_skip = calc_to_skip(page, limit)
         return list(self.commands.find({}).sort("type", ASCENDING).skip(to_skip).limit(limit)), HTTPStatus.OK
